Refactor puzzle-related constants and functions: move PUZZLE_MESSAGE_DELETE_MS to puzzle.ts, implement getPuzzleNameFormat in puzzle-utils.ts, and remove unused code from publishers.ts.
