+++ commands/puzzle.ts
@@ -1,10 +1,10 @@
 import { SlashCommandBuilder, EmbedBuilder } from "discord.js";
 import type { ChatInputCommandInteraction } from "discord.js";
 import { publisherButtons } from "../utilities/buttons";
-import {
-  EMBED_COLOR_SELECT,
-  PUZZLE_MESSAGE_DELETE_MS,
-} from "../constants/embed-colors";
+import { EMBED_COLOR_SELECT } from "../constants/embed-colors";
+
+/** Delay in ms before the /puzzle prompt message is auto-deleted. */
+const PUZZLE_MESSAGE_DELETE_MS = 15_000;
 
 function isDiscordApiError(error: unknown): error is { code: number } {
   return (
