+++ constants/embed-colors.ts
@@ -1,6 +1,3 @@
 export const EMBED_COLOR_PUZZLE = 0x78a6ee;
 export const EMBED_COLOR_SOLVED = 0x6ca24d;
 export const EMBED_COLOR_SELECT = 0x91a774;
-
-/** Delay in ms before the /puzzle prompt message is auto-deleted. */
-export const PUZZLE_MESSAGE_DELETE_MS = 15_000;
