+++ constants/publishers.ts
@@ -23,45 +23,3 @@ export const PAGE2_IDS: PublisherId[] = [
   "universal",
   "new yorker",
 ];
-
-export function getPuzzleNameFormat(
-  publisherId: PublisherId,
-  date: Date
-): string {
-  const d = String(date.getDate());
-  const dd = d.padStart(2, "0");
-  const month = date.toLocaleString("en-US", { month: "long" });
-  const monthShort = date.toLocaleString("en-US", { month: "short" });
-  const yyyy = date.getFullYear();
-  const weekdays = [
-    "Sunday",
-    "Monday",
-    "Tuesday",
-    "Wednesday",
-    "Thursday",
-    "Friday",
-    "Saturday",
-  ];
-  const weekdaysShort = ["Sun", "Mon", "Tue", "Wed", "Thu", "Fri", "Sat"];
-  const weekday = weekdays[date.getDay()];
-  const weekdayShort = weekdaysShort[date.getDay()];
-
-  switch (publisherId) {
-    case "nyt":
-      return `NY Times, ${weekday}, ${month} ${d}, ${yyyy}`;
-    case "lat":
-      return `LA Times, ${weekdayShort}, ${monthShort} ${d}, ${yyyy}`;
-    case "usa":
-      return `USA Today ${weekday}, ${monthShort} ${dd}, ${yyyy}`;
-    case "wsj":
-      return `WSJ ${weekday}, ${monthShort} ${dd}, ${yyyy}`;
-    case "newsday":
-      return `Newsday ${weekday}, ${monthShort} ${dd}, ${yyyy}`;
-    case "universal":
-      return `Universal Crossword ${weekday}`;
-    case "new yorker":
-      return `New Yorker ${weekday}, ${monthShort} ${dd}, ${yyyy}`;
-    default:
-      return "";
-  }
-}
